legger til justeringer basert på tilbakemeldinger
